feat: blocks form when showing attendance and displays attendance summaries
