+++ semis/attendance/src/main/java/org/saudigitus/semis/attendance/ui/AttendanceScreen.kt
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -52,7 +51,6 @@ import org.saudigitus.semis.attendance.ui.model.BottomSheetType
 import org.saudigitus.semis.attendance.ui.model.ButtonStep
 import org.saudigitus.semis.core.designsystem.R
 import org.saudigitus.semis.core.designsystem.attendance.model.AttendanceButtonModel
-import org.saudigitus.semis.core.designsystem.components.ConfigNotFound
 import org.saudigitus.semis.core.designsystem.components.NoResults
 import org.saudigitus.semis.core.designsystem.components.SnackBar
 import org.saudigitus.semis.core.designsystem.components.ToolbarActionState
@@ -185,7 +183,7 @@ fun AttendanceScreen(
         } else if (state.teis.isEmpty()) {
             NoResults(message = stringResource(id = R.string.no_records_found))
         } else {
-            if (state.attendanceButtonState.buttons.isEmpty() && !state.attendanceButtonState.isLoading) {
+            /*if (state.attendanceButtonState.buttons.isEmpty() && !state.attendanceButtonState.isLoading) {
                 ConfigNotFound(
                     modifier = Modifier
                         .fillMaxWidth()
@@ -194,7 +192,7 @@ fun AttendanceScreen(
                     iconSize = 32.dp,
                     message = stringResource(id = R.string.app_not_properly_config)
                 )
-            }
+            }*/
 
             LazyColumn(
                 modifier = Modifier.fillMaxSize(),
@@ -259,7 +257,6 @@ fun AttendanceScreen(
                             key = tei.uid(),
                             modifier = Modifier.fillMaxWidth(),
                             state = formState,
-                            attendanceButtonState = state.attendanceButtonState,
                             onEvent = onFormEvent
                         )
                         Spacer(modifier = Modifier.padding(2.dp))
