+++ semis/attendance/src/main/java/org/saudigitus/semis/attendance/ui/AttendanceViewModel.kt
@@ -2,7 +2,6 @@ package org.saudigitus.semis.attendance.ui
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Rocket
-import androidx.compose.ui.graphics.Color
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -14,34 +13,23 @@ import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import org.dhis2.commons.resources.ResourceManager
-import org.hisp.dhis.android.core.maintenance.D2Error
 import org.saudigitus.semis.attendance.R
-import org.saudigitus.semis.attendance.data.repository.AttendanceEventRepository
-import org.saudigitus.semis.attendance.data.repository.AttendanceOptionRepository
 import org.saudigitus.semis.attendance.ui.model.BottomSheetConfirmAction
 import org.saudigitus.semis.attendance.ui.model.BottomSheetType
 import org.saudigitus.semis.attendance.ui.model.ButtonStep
 import org.saudigitus.semis.core.data.model.SearchTeiModel
 import org.saudigitus.semis.core.data.model.app_config.Attendance
 import org.saudigitus.semis.core.data.repository.AppConfigRepository
-import org.saudigitus.semis.core.designsystem.attendance.AttendanceButtonState
-import org.saudigitus.semis.core.designsystem.attendance.model.AttendanceButtonDecorator
 import org.saudigitus.semis.core.designsystem.attendance.model.AttendanceButtonModel
-import org.saudigitus.semis.core.designsystem.attendance.model.AttendanceEvent
-import org.saudigitus.semis.core.designsystem.attendance.model.AttendanceEventWithDecorator
 import org.saudigitus.semis.core.designsystem.components.FilterDetailsState
-import org.saudigitus.semis.core.designsystem.components.bottomsheet.model.BottomSheetModel
 import org.saudigitus.semis.core.designsystem.components.model.ToolbarHeaders
-import org.saudigitus.semis.core.designsystem.theme.white
-import org.saudigitus.semis.core.designsystem.utils.UiDefaults
-import org.saudigitus.semis.core.designsystem.utils.UiDefaults.getAttendanceStatusColor
+import org.saudigitus.semis.core.form.data.repository.FormRepository
 import org.saudigitus.semis.core.utils.DateHelper
 import javax.inject.Inject
 
 @HiltViewModel
 class AttendanceViewModel @Inject constructor(
-    private val attendanceOptionRepository: AttendanceOptionRepository,
-    private val attendanceEventRepository: AttendanceEventRepository,
+    private val formRepository: FormRepository,
     private val appConfigRepository: AppConfigRepository,
     private val resourceManager: ResourceManager
 ) : ViewModel() {
@@ -81,20 +69,6 @@ class AttendanceViewModel @Inject constructor(
                 )
             }
         }
-    }
-
-    private suspend fun buttonState(
-        program: String,
-        attendanceEvents: List<AttendanceEventWithDecorator>
-    ): AttendanceButtonState {
-        val current = uiState.value.attendanceButtonState
-        val options = attendanceOptionRepository.getAttendanceStatusOptions(program)
-
-        return current.copy(
-            isLoading = false,
-            buttons = options,
-            attendanceEvents = attendanceEvents
-        )
     }
 
     fun initialize(
@@ -145,7 +119,7 @@ class AttendanceViewModel @Inject constructor(
                 )
             }
 
-            val attendanceEvents = attendanceEventRepository.getAttendanceEvent(
+            val currentButtonState = formRepository.loadAttendanceEvents(
                 teiUids = studentsIds,
                 program = uiState.value.program,
                 programStage = attendanceConfig?.programStage.orEmpty(),
@@ -155,14 +129,11 @@ class AttendanceViewModel @Inject constructor(
                     .orEmpty()
             )
 
-            val currentButtonState = buttonState(uiState.value.program, attendanceEvents)
-
             _uiState.update {
                 it.copy(
                     isLoading = false,
                     hasDataToSave = false,
                     toolbarHeaders = updatedToolbar,
-                    attendanceButtonState = currentButtonState,
                     genericsBottomSheetState = currentBulkBottomSheet.copy(
                         imageVector = Icons.Default.Rocket,
                         title = resourceManager.getString(R.string.bulk_attendance),
@@ -175,178 +146,57 @@ class AttendanceViewModel @Inject constructor(
 
     private fun attendanceSummary() {
         viewModelScope.launch {
-            val current = uiState.value.attendanceButtonState.attendanceEvents
             val attendanceSummaryState = uiState.value.attendanceSummaryState
-
-            val options =
-                attendanceOptionRepository.getAttendanceStatusOptions(uiState.value.program)
-
-            val summaries = options.map { option ->
-                val count = current.count { option.code == it.event?.value }
-
-                BottomSheetModel(
-                    icon = option.icon,
-                    iconName = option.iconName,
-                    label = option.name,
-                    value = "$count",
-                    color = option.color
-                )
-            }
-
-            _uiState.update {
-                it.copy(
-                    attendanceSummaryState = attendanceSummaryState.copy(
-                        bottomSheetModels = summaries
+            formRepository.attendanceSummary(uiState.value.program) { summaries ->
+                _uiState.update {
+                    it.copy(
+                        attendanceSummaryState = attendanceSummaryState.copy(
+                            bottomSheetModels = summaries
+                        )
                     )
-                )
+                }
             }
         }
     }
 
     private fun updateAttendanceEvent(
-        buttonState: AttendanceButtonState,
-        attendanceEvents: MutableList<AttendanceEventWithDecorator>,
         tei: SearchTeiModel?,
         buttonModel: AttendanceButtonModel
     ) {
-        val event = attendanceEvents.find {
-            it.event?.tei == tei?.uid()
-        }
-
-        if (event != null) {
-            val updatedEvent = event.event?.copy(value = buttonModel.code.orEmpty())
-            val eventWithDecorator = event.copy(
-                event = updatedEvent,
-                decorator = AttendanceButtonDecorator(
-                    buttonType = buttonModel.code.orEmpty(),
-                    containerColor = buttonModel.color ?: getAttendanceStatusColor(
-                        buttonModel.code.orEmpty()
-                    ),
-                    contentColor = white
-                ),
-            )
-
-            val hasBeenRemoved = attendanceEvents.removeIf { it.event?.tei == tei?.uid() }
-
-            if (hasBeenRemoved) {
-                attendanceEvents.add(eventWithDecorator)
-            }
-        } else {
-            attendanceEvents.add(
-                AttendanceEventWithDecorator(
-                    tei = tei,
-                    event = AttendanceEvent(
-                        tei = tei?.uid().orEmpty(),
-                        enrollment = tei?.selectedEnrollment?.uid().orEmpty(),
-                        dataElement = buttonModel.dataElement.orEmpty(),
-                        reasonDataElement = attendanceConfig?.absenceReason,
-                        reasonOfAbsence = "",
-                        value = buttonModel.code.orEmpty(),
-                        date = selectedDate ?: DateHelper.formatDate(System.currentTimeMillis())
-                            .orEmpty()
-                    ),
-                    decorator = AttendanceButtonDecorator(
-                        buttonType = buttonModel.code.orEmpty(),
-                        containerColor = buttonModel.color ?: getAttendanceStatusColor(
-                            buttonModel.code.orEmpty()
-                        ),
-                        contentColor = white
-                    ),
-                    icon = buttonModel.icon
-                        ?: UiDefaults.dynamicIcons(buttonModel.iconName.orEmpty()),
-                    iconName = buttonModel.iconName.orEmpty(),
-                    iconColor = Color.White,
-                )
-            )
-        }
-
-        _uiState.update {
-            it.copy(
-                hasDataToSave = true,
-                attendanceButtonState = buttonState.copy(
-                    attendanceEvents = attendanceEvents
-                )
+        viewModelScope.launch {
+            formRepository.updateAttendanceEvent(
+                selectedDate,
+                attendanceConfig?.absenceReason.orEmpty(),
+                "",
+                tei,
+                buttonModel
             )
         }
     }
 
     private fun bulkAttendance(buttonModel: AttendanceButtonModel) {
         viewModelScope.launch {
-            val current = uiState.value.attendanceButtonState
-            val attendanceEvents = current.attendanceEvents.toMutableList()
-
             uiState.value.teis.forEach {
-                updateAttendanceEvent(current, attendanceEvents, it, buttonModel)
+                formRepository
+                    .updateAttendanceEvent(
+                        selectedDate,
+                        attendanceConfig?.absenceReason.orEmpty(),
+                        "",
+                        it,
+                        buttonModel
+                    )
             }
         }
     }
-
-    private fun saveAttendanceEvents() {
-        viewModelScope.launch {
-            val current = uiState.value
-
-            runCatching {
-                attendanceEventRepository.save(
-                    program = current.program,
-                    programStage = attendanceConfig?.programStage.orEmpty(),
-                    attendanceEvents = current.attendanceButtonState.attendanceEvents
-                )
-            }.onSuccess {
-                val currentButtonState = uiState.value.attendanceButtonState
-                val currentAttendanceSummaryState = uiState.value.attendanceSummaryState
-
-                _uiState.update {
-                    it.copy(
-                        hasDataToSave = false,
-                        buttonStep = ButtonStep.NONE,
-                        attendanceSummaryState = currentAttendanceSummaryState.copy(
-                            enableBulk = false,
-                        ),
-                        attendanceButtonState = currentButtonState.copy(
-                            isEditing = false,
-                        )
-                    )
-                }
-                loadAttendanceEventsByDate(selectedDate)
-                _snackbarEvent.emit(resourceManager.getString(R.string.attendance_saved))
-            }.onFailure { error ->
-                val friendlyMessage = when (error) {
-                    is D2Error -> {
-                        "${error.errorCode()} – ${
-                            error.message ?: resourceManager
-                                .getString(R.string.error_unexpected)
-                        }"
-                    }
-
-                    else -> error.message ?: resourceManager
-                        .getString(R.string.error_unexpected)
-                }
-
-                _uiState.update {
-                    it.copy(
-                        hasDataToSave = true,
-                        errorMessage = friendlyMessage
-                    )
-                }
-                _uiState.update {
-                    it.copy(
-                        hasDataToSave = true,
-                        errorMessage = error.message
-                    )
-                }
-            }
-        }
-    }
-
 
     fun handleUiEvent(uiEvent: AttendanceUiEvent) {
         when (uiEvent) {
             is AttendanceUiEvent.OnDateSelect -> {
                 loadAttendanceEventsByDate(uiEvent.date)
+                attendanceSummary()
             }
 
             is AttendanceUiEvent.OnEditClicked -> {
-                val current = uiState.value.attendanceButtonState
                 val currentAttendanceSummaryState = uiState.value.attendanceSummaryState
 
                 _uiState.update {
@@ -355,20 +205,14 @@ class AttendanceViewModel @Inject constructor(
                         attendanceSummaryState = currentAttendanceSummaryState.copy(
                             enableBulk = true
                         ),
-                        attendanceButtonState = current.copy(
-                            isEditing = true,
-                        )
                     )
                 }
+                formRepository.allowFormEdition(true)
             }
 
             is AttendanceUiEvent.OnAttendanceClick -> {
-                val current = uiState.value.attendanceButtonState
-                val attendanceEvents = current.attendanceEvents.toMutableList()
 
                 updateAttendanceEvent(
-                    current,
-                    attendanceEvents,
                     uiEvent.tei,
                     uiEvent.buttonModel
                 )
@@ -405,7 +249,7 @@ class AttendanceViewModel @Inject constructor(
                         it.copy(displayBulk = false)
                     }
                 } else {
-                    saveAttendanceEvents()
+                    //TODO: IMPLEMENT ATTENDANCE DATA SAVE
                 }
             }
