+++ semis/core/form/src/main/java/org/saudigitus/semis/core/form/ui/state/FormUiState.kt
@@ -1,6 +1,7 @@
 package org.saudigitus.semis.core.form.ui.state
 
 import androidx.compose.runtime.Immutable
+import org.saudigitus.semis.core.designsystem.attendance.AttendanceButtonState
 import org.saudigitus.semis.core.designsystem.components.model.ToolbarHeaders
 import org.saudigitus.semis.core.form.data.model.FormFieldState
 
@@ -8,7 +9,9 @@ import org.saudigitus.semis.core.form.data.model.FormFieldState
 @Immutable
 data class FormUiState(
     val isLoading: Boolean = false,
+    val isEditing: Boolean = false,
     val formBuilderState: FormBuilderState = FormBuilderState(),
+    val attendanceButtonState: AttendanceButtonState = AttendanceButtonState(),
     val toolbarHeaders: ToolbarHeaders = ToolbarHeaders(""),
     val fields: List<FormFieldState> = emptyList(),
     val error: String? = null,
